feat #1155: Despliegue
- Refactor consultas oikos.
