+++ src/app/pages/salidas/form-elementos-seleccionados/form-elementos-seleccionados.component.ts
@@ -83,7 +83,7 @@ export class FormElementosSeleccionadosComponent implements OnInit {
     }
 
     const sede_ = this.Sedes.find((x) => x.Id === sede);
-    this.Actas_Recibido.getAsignacionesBySedeAndDependencia(sede_.CodigoAbreviacion, dependencia.Id).subscribe((res: any) => {
+    this.oikosHelper.getAsignacionesBySedeAndDependencia(sede_.CodigoAbreviacion, dependencia.Id).subscribe((res: any) => {
       this.UbicacionesFiltradas = res;
     });
   }
